pyproject.toml
[project]
name = "analize-data-llm"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.18",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "gigachat>=0.1.39.post1",
    "isort>=6.0.1",
    "kaggle>=1.7.4.5",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]
